ADD vision camera on UI
