+++ user-interface/src/components/QRCameraPanel.js
@@ -1,7 +1,7 @@
 import socketIOClient from "socket.io-client";
 
 const QRCameraPanel = () => {
-  const ENDPOINT = "http://10.247.208.54:4000";
+  const ENDPOINT = "http://192.168.100.7:4000";
 
   const socket = socketIOClient(ENDPOINT);
   socket.on('video',function(data){
@@ -9,15 +9,10 @@ const QRCameraPanel = () => {
     document.getElementById("myImageID").src=string_src;
   })
 
-  socket.on('close',function(data){
-    document.getElementById('myImageID').src="./logo192.png";
-  })
-
   return (
     <div className="border shadow-sm bg-body p-2 rounded h-75">
       <p className="text-start text-uppercase text-muted">
-        {/* <small>QR Camera</small> */}
-        <img id="myImageID" src="" alt="QR Camera" width="400" height="500"></img>
+        <img className="w-100 h-100" id="myImageID" src="" alt="QR Camera"></img>
       </p>
     </div>
   );
